voice/requirements.txt
# voice/requirements.txt — Pipecat voice service dependencies
#
# Core: pipecat-ai with transport/STT/TTS integrations
# Web:  FastAPI + uvicorn
# DB:   SQLAlchemy async + asyncpg
# Obs:  Prometheus metrics

# ─── Pipecat AI (voice pipeline framework) ────────────────────────────────────
pipecat-ai[daily,silero,deepgram,cartesia,openai]==0.0.54

# ─── Web framework ────────────────────────────────────────────────────────────
fastapi>=0.109.0,<0.112.0
uvicorn[standard]>=0.27.0,<0.31.0
python-multipart>=0.0.6,<0.1.0

# ─── Database (async PostgreSQL) ──────────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.25,<2.1.0
asyncpg>=0.29.0,<0.30.0

# ─── Redis (session state) ────────────────────────────────────────────────────
redis>=5.0.0,<6.0.0

# ─── HTTP client (billing API calls) ──────────────────────────────────────────
httpx>=0.27.0,<0.28.0

# ─── Configuration ────────────────────────────────────────────────────────────
pydantic-settings>=2.1.0,<2.3.0

# ─── Observability ────────────────────────────────────────────────────────────
prometheus-client>=0.19.0,<0.22.0

# ─── Twilio (webhook signature validation) ────────────────────────────────────
twilio>=9.0.0,<10.0.0

# ─── Daily.co (WebRTC transport) ──────────────────────────────────────────────
daily-python>=0.10.0,<0.12.0
